requirements.txt
biopython==1.65
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
numpy==1.9.2
SQLAlchemy==0.9.8
Werkzeug==0.10.1
WTForms==2.0.2
